Max Velocity
Added maxVelocity threshhold to limit an issue with collision detection,
previosly movements that were to fast to comprehend at 60fps, were
ignored. This resulted in objects falling through floors etc.
The maxVelocity param is only a hack and should be replaced by a
function which tracks the complete movement of objects in between
frames, not, as it is now, the state of the object in the current frame.